// src/app/model/tabla-caja.ts
export class  TbCaja {
    bl_codbco: string;
    bl_nombco: string;
    bl_tipmon: string;
    bl_numcta: string;
    bl_estado: string;
    bl_cuenta: string;
    bl_desccta: string;
    bl_entfina: string;
    bl_tipocta: string;
    bl_flcaja: string;
    bl_fche1: string;
    bl_fche2: string;
    bl_fche3: string;
    bl_nche1: string;
    bl_nche2: string;
    bl_nche3: string;
    bl_frche: string;
    bl_modesdcta: string;
    bl_usrcrea: string;
    bl_feccrea: string;
    bl_hracrea: string;
    bl_usract: string;
    bl_fecact: string;
    bl_hraact: string;
}